+++ Editor/IClassAttributeEnforcer.cs
@@ -1,4 +1,15 @@
 ﻿namespace Ikonoclast.ClassAttributes.Editor
 {
-    public interface IClassAttributeEnforcer { }
+    using Common.Editor;
+    using UnityEngine;
+
+    public interface IClassAttributeEnforcer : IEditorSaveObject
+    {
+        float ConfigurationViewHeight
+        {
+            get;
+        }
+
+        void OnConfigurationGUI(Vector2 size);
+    }
 }