+++ Editor/RequireChildComponentClassAttributeEnforcer.cs
@@ -6,6 +6,8 @@ using System.Collections.Generic;
 
 namespace Ikonoclast.ClassAttributes.Editor
 {
+    using Common;
+    using Common.Editor;
     using static ClassAttributeEnforcerUtilities.UndoGroupNames;
 
     using Utilities = ClassAttributeEnforcerUtilities<RequireChildComponentAttribute>;
@@ -18,12 +20,42 @@ namespace Ikonoclast.ClassAttributes.Editor
     {
         #region Fields
 
+        private static bool enabled = false;
         private static Assembly[] assemblies;
         private static List<Type> typesWithAttribute;
 
         #endregion
 
         #region Properties
+
+        public static bool Enabled
+        {
+            get
+            {
+                return enabled;
+            }
+            set
+            {
+                if (value == enabled)
+                    return;
+
+                if (value == true)
+                {
+                    Subscribe();
+                }
+                else
+                {
+                    Unsubscribe();
+                }
+
+                enabled = value;
+
+                ClassAttributeEnforcerUtilities.ShouldReloadConfiguration = true;
+            }
+        }
+
+        private static ISaveObject SaveObject =>
+            Instance;
 
         private static RequireChildComponentClassAttributeEnforcer Instance
         {
@@ -33,6 +65,19 @@ namespace Ikonoclast.ClassAttributes.Editor
         #endregion
 
         #region Constructors
+
+        public RequireChildComponentClassAttributeEnforcer()
+        {
+            Utilities.Register(this);
+
+            enabled = ClassAttributeEnforcerUtilities
+                .LoadConfiguration<bool>($"{nameof(RequireChildComponentClassAttributeEnforcer)}.{nameof(Enabled)}");
+
+            if (enabled)
+            {
+                Subscribe();
+            }
+        }
 
         static RequireChildComponentClassAttributeEnforcer()
         {
@@ -45,6 +90,8 @@ namespace Ikonoclast.ClassAttributes.Editor
 
         ~RequireChildComponentClassAttributeEnforcer()
         {
+            Utilities.Unregister(this);
+
             Unsubscribe();
         }
 
@@ -56,12 +103,14 @@ namespace Ikonoclast.ClassAttributes.Editor
         {
             EditorApplication.hierarchyChanged += OnHierarchyChanged;
             UnityEditor.SceneManagement.EditorSceneManager.sceneOpened += OnSceneOpened;
+            ClassAttributeEnforcerEditorWindow.RequestAttributeEnforcement += OnRequestAttributeEnforcement;
         }
 
         private static void Unsubscribe()
         {
             EditorApplication.hierarchyChanged -= OnHierarchyChanged;
             UnityEditor.SceneManagement.EditorSceneManager.sceneOpened -= OnSceneOpened;
+            ClassAttributeEnforcerEditorWindow.RequestAttributeEnforcement -= OnRequestAttributeEnforcement;
         }
 
         private static void OnProjectLoaded()
@@ -73,7 +122,11 @@ namespace Ikonoclast.ClassAttributes.Editor
 
         private static void OnHierarchyChanged()
         {
-            if (!Undo.GetCurrentGroupName().ContainsAny(AddComponent, RemoveComponent, DeleteGameObject, ClassAttributeEnforcer))
+            if (!Enabled)
+                return;
+
+            // Only apply the attribute if the last operation had a name containing one of these strings.
+            if (!Undo.GetCurrentGroupName().ContainsAny(Drag, AddComponent, RemoveComponent, DeleteGameObject, ClassAttributeEnforcer))
                 return;
 
             Utilities.ApplyAttributeToAllGameObjectsOfType(typesWithAttribute, ApplyAttribute);
@@ -86,14 +139,17 @@ namespace Ikonoclast.ClassAttributes.Editor
 
             void AddRequiredComponent(KeyValuePair<Type, string> kvp)
             {
+                // Verify that the child being added does not have the RequireNameAttribute.
                 if (Attribute.GetCustomAttribute(kvp.Key, typeof(RequireNameAttribute)) != null)
                     throw new NotSupportedException($"Conflicting attributes on {kvp.Key.FullName}. " +
-                                $"{nameof(RequireChildComponentAttribute)} cannot be used with {nameof(RequireNameAttribute)}.");
+                        $"{nameof(RequireChildComponentAttribute)} cannot be used with {nameof(RequireNameAttribute)}.");
 
-                string childName = kvp.Value;
+                var childName = kvp.Value;
 
                 var parent = gameObject.transform;
 
+                // Check every child of the game object, and
+                // add the required component if one with a matching name is found without the component.
                 foreach (Transform child in parent)
                 {
                     if (child.name == childName)
@@ -104,11 +160,12 @@ namespace Ikonoclast.ClassAttributes.Editor
                     }
                 }
 
+                // Create a new child with the specified (or default) name, add the component, and assign its parent.
                 var newChild = new GameObject(childName);
 
                 newChild.tag = parent.tag;
                 newChild.AddComponent(kvp.Key);
-                newChild.transform.parent = parent;
+                newChild.transform.SetParent(parent);
             }
 
             var required = attr.required;
@@ -128,7 +185,7 @@ namespace Ikonoclast.ClassAttributes.Editor
                     foreach (var comp in components)
                     {
                         // Check that the component is on a child gameObject.
-                        if (comp.gameObject != gameObject)  // NullRef with IActorVision attr on AIActor
+                        if (comp.gameObject != gameObject)  // TODO: Investigate null reference with interfaces
                         {
                             foundChild = true;
 
@@ -143,12 +200,78 @@ namespace Ikonoclast.ClassAttributes.Editor
                 }
             }
 
+            // Set the undo group name to avoid repeated enforcements.
             Undo.SetCurrentGroupName($"Undo $[{nameof(RequireChildComponentClassAttributeEnforcer)}]");
         }
 
         private static void OnSceneOpened(UnityEngine.SceneManagement.Scene scene, UnityEditor.SceneManagement.OpenSceneMode mode)
         {
             OnHierarchyChanged();
+        }
+
+        #endregion
+
+        #region Event Listeners
+
+        private static void OnRequestAttributeEnforcement()
+        {
+            if (Enabled)
+            {
+                Utilities.ApplyAttributeToAllGameObjectsOfType(typesWithAttribute, ApplyAttribute);
+            }
+        }
+
+        #endregion
+
+        #region IEditorSaveObject Implementations
+
+        string ISaveObject.ID =>
+            nameof(RequireChildComponentClassAttributeEnforcer);
+
+        Map ISaveObject.Serialize()
+        {
+            var map = new Map(SaveObject.ID);
+
+            map[$"{SaveObject.ID}.{nameof(Enabled)}"] = Enabled;
+
+            return map;
+        }
+
+        void ISaveObject.Serialize(Map map, bool overwrite)
+        {
+            if (!overwrite && map.HasKey($"{SaveObject.ID}.{nameof(Enabled)}"))
+                return;
+
+            map[$"{SaveObject.ID}.{nameof(Enabled)}"] = Enabled;
+        }
+
+        void ISaveObject.Deserialize(Map dict)
+        {
+            Enabled = dict.GetRawBoolean($"{SaveObject.ID}.{nameof(Enabled)}");
+        }
+
+        bool IEditorSaveObject.Enabled
+        {
+            get => Enabled;
+            set => Enabled = value;
+        }
+
+        void IEditorSaveObject.Reset()
+        {
+            Enabled = true;
+        }
+
+        #endregion
+
+        #region IClassAttributeEnforcer Implementations
+        float IClassAttributeEnforcer.ConfigurationViewHeight =>
+            Utilities.BoolConfigurationHeight;
+
+        void IClassAttributeEnforcer.OnConfigurationGUI(Vector2 size)
+        {
+            float offsetY = 0;
+
+            Enabled = Utilities.MakeBoolConfiguration(size, Enabled, nameof(Enabled), ref offsetY);
         }
 
         #endregion