+++ Editor/RequireComponentClassAttributeEnforcer.cs
@@ -6,6 +6,8 @@ using System.Collections.Generic;
 
 namespace Ikonoclast.ClassAttributes.Editor
 {
+    using Common;
+    using Common.Editor;
     using static ClassAttributeEnforcerUtilities.UndoGroupNames;
 
     using Utilities = ClassAttributeEnforcerUtilities<RequireComponentAttribute>;
@@ -19,12 +21,42 @@ namespace Ikonoclast.ClassAttributes.Editor
     {
         #region Fields
 
+        private static bool enabled = false;
         private static Assembly[] assemblies;
         private static List<Type> typesWithAttribute;
 
         #endregion
 
         #region Properties
+
+        public static bool Enabled
+        {
+            get
+            {
+                return enabled;
+            }
+            set
+            {
+                if (value == enabled)
+                    return;
+
+                if (value == true)
+                {
+                    Subscribe();
+                }
+                else
+                {
+                    Unsubscribe();
+                }
+
+                enabled = value;
+
+                ClassAttributeEnforcerUtilities.ShouldReloadConfiguration = true;
+            }
+        }
+
+        private static ISaveObject SaveObject =>
+            Instance;
 
         private static RequireComponentClassAttributeEnforcer Instance
         {
@@ -34,6 +66,19 @@ namespace Ikonoclast.ClassAttributes.Editor
         #endregion
 
         #region Constructors
+
+        public RequireComponentClassAttributeEnforcer()
+        {
+            Utilities.Register(this);
+
+            enabled = ClassAttributeEnforcerUtilities
+                .LoadConfiguration<bool>($"{nameof(RequireComponentClassAttributeEnforcer)}.{nameof(Enabled)}");
+
+            if (enabled)
+            {
+                Subscribe();
+            }
+        }
 
         static RequireComponentClassAttributeEnforcer()
         {
@@ -46,6 +91,8 @@ namespace Ikonoclast.ClassAttributes.Editor
 
         ~RequireComponentClassAttributeEnforcer()
         {
+            Utilities.Unregister(this);
+
             Unsubscribe();
         }
 
@@ -57,12 +104,14 @@ namespace Ikonoclast.ClassAttributes.Editor
         {
             EditorApplication.hierarchyChanged += OnHierarchyChanged;
             UnityEditor.SceneManagement.EditorSceneManager.sceneOpened += OnSceneOpened;
+            ClassAttributeEnforcerEditorWindow.RequestAttributeEnforcement += OnRequestAttributeEnforcement;
         }
 
         private static void Unsubscribe()
         {
             EditorApplication.hierarchyChanged -= OnHierarchyChanged;
             UnityEditor.SceneManagement.EditorSceneManager.sceneOpened -= OnSceneOpened;
+            ClassAttributeEnforcerEditorWindow.RequestAttributeEnforcement -= OnRequestAttributeEnforcement;
         }
 
         private static void OnProjectLoaded()
@@ -74,7 +123,11 @@ namespace Ikonoclast.ClassAttributes.Editor
 
         private static void OnHierarchyChanged()
         {
-            if (!Undo.GetCurrentGroupName().ContainsAny(AddComponent, RemoveComponent, InspectorChange, ClassAttributeEnforcer))
+            if (!Enabled)
+                return;
+
+            // Only apply the attribute if the last operation had a name containing one of these strings.
+            if (!Undo.GetCurrentGroupName().ContainsAny(Drag, AddComponent, RemoveComponent, InspectorChange, ClassAttributeEnforcer))
                 return;
 
             Utilities.ApplyAttributeToSelectedGameObject(typesWithAttribute, ApplyAttribute);
@@ -87,6 +140,7 @@ namespace Ikonoclast.ClassAttributes.Editor
 
             var required = attr.required;
 
+            // Add each required component to the game object if not present.
             foreach (var type in required)
             {
                 if (!gameObject.TryGetComponent(type, out var _))
@@ -103,6 +157,7 @@ namespace Ikonoclast.ClassAttributes.Editor
                 }
             }
 
+            // Set the undo group name to avoid repeated enforcements.
             Undo.SetCurrentGroupName($"Undo $[{nameof(RequireComponentClassAttributeEnforcer)}]");
         }
 
@@ -110,6 +165,73 @@ namespace Ikonoclast.ClassAttributes.Editor
         {
             OnHierarchyChanged();
         }
+
+        #endregion
+
+        #region Event Listeners
+
+        private static void OnRequestAttributeEnforcement()
+        {
+            if (Enabled)
+            {
+                Utilities.ApplyAttributeToAllGameObjectsOfType(typesWithAttribute, ApplyAttribute);
+            }
+        }
+
+        #endregion
+
+        #region IEditorSaveObject Implementations
+
+        string ISaveObject.ID =>
+            nameof(RequireComponentClassAttributeEnforcer);
+
+        Map ISaveObject.Serialize()
+        {
+            var map = new Map(SaveObject.ID);
+
+            map[$"{SaveObject.ID}.{nameof(Enabled)}"] = Enabled;
+
+            return map;
+        }
+
+        void ISaveObject.Serialize(Map map, bool overwrite)
+        {
+            if (!overwrite && map.HasKey($"{SaveObject.ID}.{nameof(Enabled)}"))
+                return;
+
+            map[$"{SaveObject.ID}.{nameof(Enabled)}"] = Enabled;
+        }
+
+        void ISaveObject.Deserialize(Map map)
+        {
+            Enabled = map.GetRawBoolean($"{SaveObject.ID}.{nameof(Enabled)}");
+        }
+
+        bool IEditorSaveObject.Enabled
+        {
+            get => Enabled;
+            set => Enabled = value;
+        }
+
+        void IEditorSaveObject.Reset()
+        {
+            Enabled = true;
+        }
+
+        #endregion
+
+        #region IClassAttributeEnforcer Implementations
+
+        float IClassAttributeEnforcer.ConfigurationViewHeight =>
+            Utilities.BoolConfigurationHeight;
+
+        void IClassAttributeEnforcer.OnConfigurationGUI(Vector2 size)
+        {
+            float offsetY = 0;
+
+            Enabled = Utilities.MakeBoolConfiguration(size, Enabled, nameof(Enabled), ref offsetY);
+        }
+
 
         #endregion
     }